Migrate scoring APIs to double precision for v6.0.0 (#32)
* feat!: migrate scoring APIs to double precision for v6.0.0

* perf: optimize pattern-match vector construction

  - add bulk character population and ASCII fast paths
  - lazily allocate non-ASCII storage
  - rename benchmark labels to Raffinert.FuzzySharp
  - update v6 migration documentation and PR summary

* cleanup

+++ FuzzySharp.Benchmarks/PartialRatioBenchmarks.cs
@@ -8,13 +8,13 @@ namespace Raffinert.FuzzySharp.Benchmarks;
 public class PartialRatioBenchmarks
 {
     [Benchmark]
-    public int PartialRatio()
+    public double PartialRatio()
     {
         return Fuzz.PartialRatio("similar", "somewhresimlrbetweenthisstring");
     }
 
     [Benchmark]
-    public int PartialRatioClassic()
+    public double PartialRatioClassic()
     {
         return Classic.Fuzz.PartialRatio("similar", "somewhresimlrbetweenthisstring");
     }
