+++ FuzzySharp/SimilarityRatio/Scorer/Generic/CachedScorerBase.cs
@@ -4,5 +4,5 @@ namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.Generic;
 
 public abstract class CachedScorerBase<T> : ICachedRatioScorer<T> where T : IEquatable<T>
 {
-    public abstract int Score(T[] input2);
+    public abstract double Score(T[] input2);
 }
