+++ FuzzySharp/SimilarityRatio/Scorer/Generic/IRatioScorer.cs
@@ -4,10 +4,10 @@ namespace Raffinert.FuzzySharp.SimilarityRatio.Scorer.Generic;
 
 public interface IRatioScorer<in T> where T : IEquatable<T>
 {
-    int Score(T[] input1, T[] input2);
+    double Score(T[] input1, T[] input2);
 }
 
 public interface ICachedRatioScorer<in T> where T : IEquatable<T>
 {
-    int Score(T[] input2);
+    double Score(T[] input2);
 }
